+++ docs/whats-new.rst
@@ -8,6 +8,9 @@ v0.x (unreleased)
 Enhancements
 ~~~~~~~~~~~~
 
+- Salem can now plot on cartopy's maps thanks to a new
+  :py:func:`~salem.proj_to_cartopy` function.
+- Doc improvements
 - New diagnostic variable: 'WS'